+++ ProdManager/helpers/resource.py
@@ -1,7 +1,10 @@
 from flask import current_app
 from sqlalchemy.exc import IntegrityError
 from ProdManager import db
-from ProdManager.helpers.response import NotFoundError, ServerError, ConflictError, UndeletableRessourceError, DependencyError
+from ProdManager.helpers.response import (
+  NotFoundError, ServerError, ConflictError,
+  UndeletableRessourceError, DependencyError,
+)
 
 def list_resources(ressource_class):
   # https://flask-sqlalchemy.palletsprojects.com/en/2.x/api/#flask_sqlalchemy.BaseQuery.paginate
@@ -31,12 +34,14 @@ def update_resource(resource_class, ressource_id, attributs):
   if resource is None:
     raise NotFoundError(ressource_id)
 
-  changed = False
+  changed = {}
 
-  for attribute, value in attributs.items():
-    if getattr(resource, attribute) != value:
-      setattr(resource, attribute, value)
-      changed = True
+  for attribute, new_value in attributs.items():
+    old_value = getattr(resource, attribute)
+
+    if old_value != new_value:
+      setattr(resource, attribute, new_value)
+      changed[attribute] = (old_value, new_value)
 
   try:
     db.session.commit()